# public/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "resend-conversation-service API",
    "version": "0.8.0",
    "description": "Receives signed Resend webhooks and provides bearer-authenticated topic-centered conversations plus synchronous and queued direct email. Conversation API v2 is the only conversation contract; conversation API v1 was retired in 0.5.0 and its paths are no longer routed. The service persists outbound intent before contacting Resend, preserves RFC reply relationships for conversations, and stores untrusted email content without retrieving attachments. The separate AsyncAPI contract for conversation lifecycle events is available from the deployed service at `/asyncapi.json`.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current application host"
    }
  ],
  "security": [
    {
      "emailV2Auth": []
    }
  ],
...
  "paths": {
...
    "/api/webhooks/resend/v1": {
      "post": {
        "operationId": "receiveResendWebhook",
        "summary": "Receive a Resend webhook",
        "description": "Verifies the Svix signature against the exact raw request body, classifies the event, and stores its normalized fields. For email.received, the service retrieves the text, HTML, and headers from Resend and updates the conversation projection; attachments are not fetched. Eligible RFC Message-ID ancestry remains authoritative. When RFC ancestry does not resolve a conversation, a recognized conversation token in to or received_for is used only when the sender matches that conversation's participant. Reconciliation of out-of-order rows preserves one-way V2 promotion and historical routing-token/base aliases when conversations merge. All three Svix headers are required and must describe the submitted body. A delivery using an svix-id already stored for the same event family is acknowledged with the same 200 response after projection succeeds. Resend may retry non-200 responses. Swagger UI cannot generate valid Svix signatures; interactive requests require header values generated for the exact raw request body.",
        "tags": ["Resend webhooks"],
        "security": [
          {
            "svixId": [],
            "svixTimestamp": [],
            "svixSignature": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "A supported Resend event serialized as JSON. The serialized bytes must be the same bytes used to generate svix-signature.",
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/EmailWebhookEvent"
                  },
                  {
                    "$ref": "#/components/schemas/ContactWebhookEvent"
                  },
                  {
                    "$ref": "#/components/schemas/DomainWebhookEvent"
                  }
                ]
              },
              "examples": {
                "emailSent": {
                  "summary": "Email accepted for delivery",
                  "value": {
                    "type": "email.sent",
                    "created_at": "2026-07-18T12:00:00.000Z",
                    "data": {
                      "email_id": "em_test123",
                      "from": "sender@example.com",
                      "to": ["recipient@example.com"],
                      "subject": "Test email",
                      "created_at": "2026-07-18T11:59:59.000Z",
                      "tags": [
                        {
                          "name": "campaign",
                          "value": "test"
                        }
                      ]
                    }
                  }
                },
                "contactCreated": {
                  "summary": "Contact created",
                  "value": {
                    "type": "contact.created",
                    "created_at": "2026-07-18T12:00:00.000Z",
                    "data": {
                      "id": "ct_test123",
                      "audience_id": "aud_test123",
                      "segment_ids": ["seg_1", "seg_2"],
                      "created_at": "2026-07-18T11:59:58.000Z",
                      "updated_at": "2026-07-18T11:59:58.000Z",
                      "email": "contact@example.com",
                      "first_name": "John",
                      "last_name": "Doe",
                      "unsubscribed": false
                    }
                  }
                },
                "domainCreated": {
                  "summary": "Domain created",
                  "value": {
                    "type": "domain.created",
                    "created_at": "2026-07-18T12:00:00.000Z",
                    "data": {
                      "id": "dom_test123",
                      "name": "example.com",
                      "status": "pending",
                      "created_at": "2026-07-18T11:59:57.000Z",
                      "region": "us-east-1",
                      "records": [
                        {
                          "record": "SPF",
                          "name": "example.com",
                          "type": "TXT",
                          "value": "v=spf1 include:_spf.resend.com ~all",
                          "ttl": "3600",
                          "status": "not_started"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The event was stored, or an event with the same svix-id was already stored for its event family.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookAcceptedResponse"
                },
                "example": {
                  "received": true
                }
              }
            }
          },
          "400": {
            "description": "One or more required Svix headers are missing, or the signed event type does not belong to a recognized Resend event family.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingHeaders": {
                    "value": {
                      "error": "Missing required Svix headers"
                    }
                  },
                  "unknownEvent": {
                    "value": {
                      "error": "Unknown event type"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "The Svix signature could not be verified against the exact raw request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid webhook signature"
                }
              }
            }
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedContentEncoding"
          },
          "500": {
            "description": "The server is missing required configuration, could not retrieve inbound content, or could not persist the event and projection. Resend can retry the delivery.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "configuration": {
                    "value": {
                      "error": "Server misconfiguration"
                    }
                  },
                  "processing": {
                    "value": {
                      "error": "Failed to process webhook"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/emails/v2": {
      "post": {
        "operationId": "sendDirectEmailV2",
        "summary": "Send a direct email without a conversation",
        "description": "Persists a direct outbound email intent and synchronously sends it through Resend. The normalized from.address must exactly match a FROM allowlist entry; recipients are not allowlisted. No conversation, parent, routing token, Reply-To header, or RFC threading headers are created. Without a Reply-To header, mail clients normally direct replies to the From address.",
        "tags": ["Direct email"],
        "security": [
          {
            "emailV2Auth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectEmailV2Request"
              },
              "example": {
                "from": {
                  "address": "notifications@example.com",
                  "name": "Example"
                },
                "to": [
                  {
                    "address": "person@example.com",
                    "name": "Person"
                  },
                  {
                    "address": "second@example.com",
                    "name": "Second Person"
                  }
                ],
                "subject": "Verify your email",
                "text": "Use the verification link.",
                "html": "<p>Use the verification link.</p>",
                "tags": [
                  {
                    "name": "category",
                    "value": "confirm_email"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A completed idempotent replay.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectEmailSendResult"
                }
              }
            }
          },
          "201": {
            "description": "Resend accepted the newly persisted email.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectEmailSendResult"
                },
                "example": {
                  "email": {
                    "id": "0198409b-7c01-7def-8ad2-8b94ad8e8987",
                    "state": "accepted",
                    "resendEmailId": "4f8f0f13-8d5a-4e79-b32a-70f39d0a3f18"
                  }
                }
              }
            }
          },
          "202": {
            "description": "An idempotent replay found intent that is still pending.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectEmailSendResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedContentEncoding"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "description": "The persisted send failed or its provider outcome is indeterminate.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectEmailSendFailure"
                }
              }
            }
          }
        }
      }
    },
    "/api/emails/v2/outbox": {
      "post": {
        "operationId": "enqueueDirectEmailV2",
        "summary": "Enqueue a direct email without a conversation",
        "description": "Atomically persists a pending direct email and shared outbox entry without calling Resend. Validation, multiple-recipient support, tags, exact FROM authorization, and Reply-To rejection match the synchronous direct-email operation. Later allowlist removal does not cancel already-persisted intent. Replaying the same key and normalized queued request returns 202 while pending, 200 after acceptance, or 502 after a failed or indeterminate drain outcome.",
        "tags": ["Direct email", "Outbox"],
        "security": [
          {
            "emailV2Auth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectEmailV2Request"
              },
              "example": {
                "from": {
                  "address": "notifications@example.com",
                  "name": "Example"
                },
                "to": [
                  {
                    "address": "person@example.com",
                    "name": "Person"
                  }
                ],
                "subject": "Verify your email",
                "html": "<p>Use the verification link.</p>",
                "tags": [
                  {
                    "name": "category",
                    "value": "verification"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A matching queued request was already accepted by Resend.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectEmailSendResult"
                }
              }
            }
          },
          "202": {
            "description": "The direct email is durably queued and remains pending.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectEmailSendResult"
                },
                "example": {
                  "email": {
                    "id": "0198409b-7c01-7def-8ad2-8b94ad8e8987",
                    "state": "pending",
                    "resendEmailId": null
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedContentEncoding"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "502": {
            "description": "The queued email was later marked failed or indeterminate.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectEmailSendFailure"
                }
              }
            }
          }
        }
      }
    },
    "/api/emails/v2/outbox/drain": {
      "post": {
        "operationId": "drainEmailOutboxV2",
        "summary": "Deliver one shared outbox batch",
        "description": "Delivers one bounded, ordered batch from the shared transactional outbox. It uses only OUTBOX_DRAIN_API_KEY and may deliver direct and conversation messages in one fixed ordered batch. This is the only drain route; invoke it at least once per minute when using queued sends. The service can also drive this same drain internally on a cron schedule instead of an external caller; that scheduler is disabled by default and is configured with OUTBOX_DRAIN_SCHEDULE_ENABLED and OUTBOX_DRAIN_SCHEDULE. A retryable provider failure preserves the fixed batch and provider idempotency key for a later invocation inside the safety window. Final failed or indeterminate outcomes are terminal and are not retried.",
        "tags": ["Outbox"],
        "security": [
          {
            "outboxDrainAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DrainOutboxRequest"
              },
              "example": {
                "limit": 100
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The bounded shared drain invocation completed, including persisted retry scheduling when applicable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrainOutboxResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedContentEncoding"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/conversations/v2/summary": {
      "get": {
        "operationId": "getConversationStateSummaryV2",
        "summary": "Summarize conversations by state",
        "description": "Returns conversation counts for every state plus a paginated list of conversations in the selected states, ordered by oldest state change first so state=awaiting_us reads as a reply queue. Items carry conversation metadata only; read messages from the conversation endpoint. Reads are not version-filtered.",
        "tags": ["Conversations"],
        "security": [
          {
            "emailV2Auth": []
          }
        ],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "description": "Restrict the list to these states. Repeat the parameter or pass a comma-separated list; values are matched case-insensitively. Omit to list every state. The counts object always covers all four states regardless of this filter.",
            "explode": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ConversationState"
              }
            }
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "name": "before",
            "in": "query",
            "description": "Opaque cursor returned by the previous state-summary page.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conversation counts per state and the selected page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationStateSummary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "emailV2Auth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Current V2 credential supplied from EMAIL_v2_API_KEY or the EMAIL_V2_API_KEY fallback. EMAIL_v2_API_KEY takes precedence when both are nonempty. It authorizes conversation V2 operations and direct email; it is not interchangeable with the outbox-drain credential."
      },
      "outboxDrainAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Dedicated credential for the private scheduled outbox drain caller."
      },
      "svixId": {
        "type": "apiKey",
        "in": "header",
        "name": "svix-id",
        "description": "Unique Svix message identifier used for signature verification and idempotency."
      },
      "svixTimestamp": {
        "type": "apiKey",
        "in": "header",
        "name": "svix-timestamp",
        "description": "Svix delivery timestamp included in signature verification."
      },
      "svixSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "svix-signature",
        "description": "Svix signature generated for the exact raw request body."
      }
    },
...
  }
}