# src/server.ts
import 'dotenv/config';
import { createServer } from 'node:http';
import path from 'node:path';
import { Readable } from 'node:stream';
import { pipeline } from 'node:stream/promises';
import express, {
  type ErrorRequestHandler,
  type Request,
  type RequestHandler,
} from 'express';
import swaggerUiDist from 'swagger-ui-dist';
import { authorizeEmailV2, authorizeOutboxDrain } from '@/lib/api';
import {
  attachmentsEnabled,
  getConfiguredAttachmentStorage,
  resolveAttachmentLimits,
  startAttachmentIngestRuntime,
  startAttachmentReaperRuntime,
  stopAttachmentIngestRuntime,
  stopAttachmentReaperRuntime,
} from '@/lib/attachments';
import {
  startConversationEventRuntime,
  stopConversationEventRuntime,
} from '@/lib/conversation-event-runtime';
import { getPrismaClient } from '@/lib/database';
import { logEvent, runWithRequestContext } from '@/lib/logger';
import {
  startOutboxDrainScheduler,
  stopOutboxDrainScheduler,
} from '@/lib/outbox-drain-scheduler';
import {
  initializeTelemetry,
  shutdownTelemetry,
  telemetryEnabled,
} from '@/lib/telemetry';
import {
  changeHttpActive,
  recordHttpRequest,
  registerDatabaseGauges,
} from '@/lib/telemetry-metrics';
import { GET as downloadAttachmentV2 } from '@/routes/attachments/v2/[attachmentId]/route';
import { POST as uploadAttachmentV2 } from '@/routes/attachments/v2/route';
import { POST as enqueueMessageV2 } from '@/routes/conversations/v2/[conversationId]/messages/outbox/route';
import { POST as sendMessageV2 } from '@/routes/conversations/v2/[conversationId]/messages/route';
import {
  DELETE as deleteConversationV2,
  GET as getConversationV2,
  PATCH as patchConversationV2,
} from '@/routes/conversations/v2/[conversationId]/route';
import { POST as setConversationStateV2 } from '@/routes/conversations/v2/[conversationId]/state/route';
import { POST as enqueueConversationV2 } from '@/routes/conversations/v2/outbox/route';
import {
  POST as createConversationV2,
  GET as listConversationsV2,
} from '@/routes/conversations/v2/route';
import { GET as getConversationSummaryV2 } from '@/routes/conversations/v2/summary/route';
import { GET as getConversationByTopicV2 } from '@/routes/conversations/v2/topics/[topicType]/[externalTopicId]/route';
import { POST as drainEmailOutboxV2 } from '@/routes/emails/v2/outbox/drain/route';
import { POST as enqueueDirectEmailV2 } from '@/routes/emails/v2/outbox/route';
import { POST as sendDirectEmailV2 } from '@/routes/emails/v2/route';
import { GET as healthV2 } from '@/routes/health/v2/route';
import { POST as webhook } from '@/routes/webhooks/resend/v1/route';

const BODY_LIMIT = '2100kb';
const rawBody = express.raw({
  type: 'application/json',
  limit: BODY_LIMIT,
  inflate: false,
});

function authorizationRequest(request: Request): globalThis.Request {
  return new globalThis.Request('http://localhost/', {
    headers: { authorization: request.get('authorization') ?? '' },
  });
}

const requireEmailV2Auth: RequestHandler = (request, response, next) => {
  const rejected = authorizeEmailV2(authorizationRequest(request));
  if (rejected) {
    rejected.headers.forEach((value, name) => {
      response.setHeader(name, value);
    });
    void rejected
      .text()
      .then((body) => response.status(rejected.status).send(body));
    return;
  }
  next();
};
const requireDrainAuth: RequestHandler = (request, response, next) => {
  const rejected = authorizeOutboxDrain(authorizationRequest(request));
  if (rejected) {
    rejected.headers.forEach((value, name) => {
      response.setHeader(name, value);
    });
    void rejected
      .text()
      .then((body) => response.status(rejected.status).send(body));
    return;
  }
  next();
};

const requireIdempotency: RequestHandler = (request, response, next) => {
  const key = request.get('idempotency-key');
  if (!key || key.length > 256) {
    response
      .status(400)
      .json({ error: 'A valid Idempotency-Key header is required' });
    return;
  }
  next();
};

function toFetchRequest(request: Request): globalThis.Request {
  const url = `${request.protocol}://${request.get('host') ?? 'localhost'}${request.originalUrl}`;
  const body = Buffer.isBuffer(request.body) ? request.body : undefined;
  return new globalThis.Request(url, {
    method: request.method,
    headers: new Headers(request.headers as Record<string, string>),
    ...(body && body.length > 0 ? { body } : {}),
  });
}

type Route = (
  request: globalThis.Request,
  context: any,
) => Promise<globalThis.Response>;
function adapt(handler: Route): RequestHandler {
  return async (request, response, next) => {
    try {
      const result = await handler(toFetchRequest(request), {
        params: Promise.resolve(request.params as Record<string, string>),
      });
      result.headers.forEach((value, name) => {
        response.setHeader(name, value);
      });
      response
        .status(result.status)
        .send(Buffer.from(await result.arrayBuffer()));
    } catch (error) {
      next(error);
    }
  };
}

/**
 * Like adapt(), but pipes the response body instead of buffering it. Used for
 * attachment downloads, which can be tens of megabytes.
 */
function adaptStream(handler: Route): RequestHandler {
  return async (request, response, next) => {
    try {
      const result = await handler(toFetchRequest(request), {
        params: Promise.resolve(request.params as Record<string, string>),
      });
      result.headers.forEach((value, name) => {
        response.setHeader(name, value);
      });
      response.status(result.status);
      if (!result.body) {
        response.end();
        return;
      }
      await pipeline(Readable.fromWeb(result.body as never), response);
    } catch (error) {
      if (response.headersSent) {
        response.destroy();
        return;
      }
      next(error);
    }
  };
}

export function createApp() {
  const app = express();
  app.disable('x-powered-by');
  app.use((request, response, next) => {
    runWithRequestContext(() => {
      const startedAt = performance.now();
      const activeAttributes = {
        'http.request.method': knownHttpMethod(request.method),
        'url.scheme': 'http',
      };
      changeHttpActive(1, activeAttributes);
      response.once('finish', () => {
        changeHttpActive(-1, activeAttributes);
        const route = matchedRoute(request);
        const durationSeconds = (performance.now() - startedAt) / 1_000;
        const attributes = {
          ...activeAttributes,
          'http.route': route,
          'http.response.status_code': response.statusCode,
        };
        recordHttpRequest(durationSeconds, attributes);
        logEvent('info', 'http_request_completed', {
          method: request.method,
          route,
          status_code: response.statusCode,
          duration_ms: Math.round(durationSeconds * 1_000),
        });
      });
      next();
    });
  });

  app.get('/api/health/v2', adapt(healthV2));
  app.post('/api/webhooks/resend/v1', rawBody, adapt(webhook));
  app.post(
    '/api/emails/v2/outbox/drain',
    requireDrainAuth,
    rawBody,
    adapt(drainEmailOutboxV2),
  );
  app.post(
    '/api/emails/v2/outbox',
    requireEmailV2Auth,
    requireIdempotency,
    rawBody,
    adapt(enqueueDirectEmailV2),
  );
  app.all('/api/emails/v2/outbox', requireEmailV2Auth, (_request, response) =>
    response.status(404).json({ error: 'Not found' }),
  );
  app.post(
    '/api/emails/v2',
    requireEmailV2Auth,
    requireIdempotency,
    rawBody,
    adapt(sendDirectEmailV2),
  );

  // Registered only when the feature is on, so with it off these paths fall
  // through to the terminal 404 exactly like the other retired surfaces.
  // Authentication runs before the body parser so an unauthenticated caller
  // can never make the process buffer a multi-megabyte upload.
  if (attachmentsEnabled()) {
    const attachmentBody = express.raw({
      type: () => true,
      limit: resolveAttachmentLimits().maxBytes,
      inflate: false,
    });
    app.post(
      '/api/attachments/v2',
      requireEmailV2Auth,
      attachmentBody,
      adapt(uploadAttachmentV2),
    );
    app.get(
      '/api/attachments/v2/:attachmentId',
      requireEmailV2Auth,
      adaptStream(downloadAttachmentV2),
    );
  }

  // Static conversation routes must precede /:conversationId routes.
  app.post(
    '/api/conversations/v2/outbox',
    requireEmailV2Auth,
    requireIdempotency,
    rawBody,
    adapt(enqueueConversationV2),
  );
  app.all(
    '/api/conversations/v2/outbox',
    requireEmailV2Auth,
    (_request, response) => response.status(404).json({ error: 'Not found' }),
  );
  app.get(
    '/api/conversations/v2/summary',
    requireEmailV2Auth,
    adapt(getConversationSummaryV2),
  );
  app.get(
    '/api/conversations/v2/topics/:topicType/:externalTopicId',
    requireEmailV2Auth,
    adapt(getConversationByTopicV2),
  );
  app
    .route('/api/conversations/v2')
    .get(requireEmailV2Auth, adapt(listConversationsV2))
    .post(
      requireEmailV2Auth,
      requireIdempotency,
      rawBody,
      adapt(createConversationV2),
    );
  app.post(
    '/api/conversations/v2/:conversationId/messages/outbox',
    requireEmailV2Auth,
    requireIdempotency,
    rawBody,
    adapt(enqueueMessageV2),
  );
  app.post(
    '/api/conversations/v2/:conversationId/messages',
    requireEmailV2Auth,
    requireIdempotency,
    rawBody,
    adapt(sendMessageV2),
  );
  app.post(
    '/api/conversations/v2/:conversationId/state',
    requireEmailV2Auth,
    rawBody,
    adapt(setConversationStateV2),
  );
  app
    .route('/api/conversations/v2/:conversationId')
    .get(requireEmailV2Auth, adapt(getConversationV2))
    .patch(requireEmailV2Auth, rawBody, adapt(patchConversationV2))
    .delete(requireEmailV2Auth, adapt(deleteConversationV2));

  app.get('/openapi.json', (_request, response, next) => {
    response.sendFile(
      path.resolve('public/openapi.json'),
      (error) => error && next(error),
    );
  });
  app.get('/asyncapi.json', (_request, response, next) => {
    response.sendFile(
      path.resolve('public/asyncapi.json'),
      (error) => error && next(error),
    );
  });
  app.use(
    '/docs/assets',
    express.static(swaggerUiDist.getAbsoluteFSPath(), { fallthrough: false }),
  );
  app.get('/docs', (_request, response) =>
    response.type('html').send(`<!doctype html>
<html lang="en"><head><meta charset="utf-8"><meta name="viewport" content="width=device-width,initial-scale=1">
<title>API documentation | resend-conversation-service</title><link rel="stylesheet" href="/docs/assets/swagger-ui.css"></head>
<body><div id="swagger-ui"></div><script src="/docs/assets/swagger-ui-bundle.js"></script>
<script>SwaggerUIBundle({url:'/openapi.json',dom_id:'#swagger-ui',deepLinking:true});</script></body></html>`),
  );

  app.use((_request, response) =>
    response.status(404).json({ error: 'Not found' }),
  );
  const errors: ErrorRequestHandler = (error, _request, response, _next) => {
    if (
      typeof error === 'object' &&
      error !== null &&
      'type' in error &&
      error.type === 'encoding.unsupported'
    ) {
      response
        .status(415)
        .json({ error: 'Compressed request bodies are not supported' });
      return;
    }
    if (
      error instanceof SyntaxError ||
      (typeof error === 'object' &&
        error !== null &&
        'type' in error &&
        error.type === 'entity.parse.failed')
    ) {
      response.status(400).json({ error: 'Request body must be valid JSON' });
      return;
    }
    if (
      typeof error === 'object' &&
      error !== null &&
      'type' in error &&
      error.type === 'entity.too.large'
    ) {
      response.status(413).json({ error: 'Request body is too large' });
      return;
    }
    logEvent('error', 'http_request_processing_failed');
    response.status(500).json({ error: 'Internal server error' });
  };
  app.use(errors);
  return app;
}

if (process.env.NODE_ENV !== 'test') {
  void startServer();
}

async function startServer() {
  try {
    initializeTelemetry();
    const client = getPrismaClient();
    if (telemetryEnabled()) {
      registerDatabaseGauges(client);
    }
    await startConversationEventRuntime(client);
    if (attachmentsEnabled()) {
      // Fail fast on unusable storage rather than accepting mail the service
      // cannot durably keep, mirroring the conversation event stream check.
      await getConfiguredAttachmentStorage().headBucket();
      logEvent('info', 'attachment_storage_verified');
    }
    startAttachmentIngestRuntime(client);
    startAttachmentReaperRuntime(client);
    startOutboxDrainScheduler(client);
    const port = Number(process.env.PORT ?? 3000);
    const host = process.env.HOST ?? process.env.HOSTNAME ?? '0.0.0.0';
    const server = createServer(createApp()).listen(port, host, () => {
      logEvent('info', 'application_listening', { port });
    });
    let closing = false;
    const shutdown = () => {
      if (closing) {
        return;
      }
      closing = true;
      logEvent('info', 'application_shutdown_started');
      server.close(async () => {
        await stopOutboxDrainScheduler();
        await stopAttachmentIngestRuntime();
        await stopAttachmentReaperRuntime();
        await stopConversationEventRuntime();
        await client.$disconnect();
        logEvent('info', 'application_shutdown_completed');
        await shutdownTelemetry();
        process.exit(0);
      });
      setTimeout(() => process.exit(1), 10_000).unref();
    };
    process.on('SIGTERM', shutdown);
    process.on('SIGINT', shutdown);
  } catch (error) {
    logEvent('error', 'application_startup_failed', {
      error_type: error instanceof Error ? error.name : 'unknown_error',
    });
    await shutdownTelemetry();
    process.exit(1);
  }
}

function matchedRoute(request: Request) {
  const route = request.route?.path;
  if (typeof route !== 'string') {
    return '/unmatched';
  }
  return `${request.baseUrl}${route}` || '/';
}

function knownHttpMethod(method: string) {
  return [
    'CONNECT',
    'DELETE',
    'GET',
    'HEAD',
    'OPTIONS',
    'PATCH',
    'POST',
    'PUT',
    'QUERY',
    'TRACE',
  ].includes(method)
    ? method
    : '_OTHER';
}